Adding IP lookup to static catalog
