+++ wrangler-core/pom.xml
@@ -135,6 +135,11 @@
       <artifactId>simmetrics-core</artifactId>
       <version>4.1.1</version>
     </dependency>
+    <dependency>
+      <groupId>com.maxmind.db</groupId>
+      <artifactId>maxmind-db</artifactId>
+      <version>1.2.1</version>
+    </dependency>
   </dependencies>
 
 </project>
